Fix bin file for valid 'bush' command; remove circular dependencies
